ChatLoop.java: Add support for tasks
Zell chatbot is lacking features.
Adding support for tasks is a good feature for a chatbot.
It will allow the user to store and manage tasks.

Let's,
* add a default chatbot command to add tasks using
  TaskList:addTask
* add a new chatbot command to print all tasks by
  printing TaskList:listAllTasks

Using a class to store and manage the tasks is the best
since managing tasks does not belong in ChatLoop.
Abstracting this to another class will also reduce
unnecessary clutter in ChatLoop.