ChatLoop.java: Shift switch case for user commands to new method
Currently handling the user commands is done using a switch case
that is placed in run(). As more commands are added it will clutter
run().

Let's, shift it to a method called handleCommand() and also change
the local variable's userInput and isFinished in run() to instance
variables. So that they can be used in handleCommand().

Abstracting the logic to handle user's commands to a function is
a good practice here. Since it will improve readibility for run()
and maintainability for handling user's commands.