ChatLoop.java: Refactor and rename handleCommand()
Currently handleCommand() does not take in any parameters but
instead uses the instance field userInput set in run(). However,
this field is only used in handleCommand().

Let's,
* rename handleCommand() to handleUserInput()
* remove the instance field userInput from the class
* change all usages of the instance field userInput to the local
  variable userInput instead

Changing userInput from an instance field to a local field is a good
practice here. Since it is pointless to have an instance field just
for it to be used in one method. Passing it as a parameter in
handleUserInput is better since it improves readibility as we can
see the dependency. In the future should it be used in multiple
methods we can reconsider changing it back to an instance field.