UserInputHandler.java: Shift over input handling logic from ChatLoop.java
Currently handling the user inputs is done in
ChatLoop:handleUserInput(). However, with the addition of new commands
and handling of exceptions that method is getting to complex.

Let's,
* Shift all the logic to handle user inputs to UserInputHandler.java
* Create a method handleInput for handling user inputs
* Create a method handleCommand to determine what to do with commands
* Create a method handleTaskCommands to handle commands under Task like
  todo, deadline, event
* Create a method createTask to handle creation of task
* Create a method getEndProgram so that ChatLoop knows if based on
  the current user input if the ChatLoop should terminate
* Create a method handleBye to handle the bye command
* Create a method handleList to handle the list command
* Create a method handleMarkOrUnMark to handle mark/unmark command
* Create all the respective methods to check and throw exceptions for
  each command.

Abstracting the logic to handle user's commands to a class a good practice
here. Since it will improve maintainability for handling user inputs in
the future since it will continue to get even more complex when more
requirements are added. Splitting the methods as stated above is good so
that we can adhere to SLAP.