Rework commands and command methods
The current code structure for commands is rather rigid, with each
command having its own parsing and execution process. This makes it
tedious to implement different variants of the same command, such as
list by date, list by id and list by keyword.

To fix this, a separate function findTasks serves the role of parsing
modifiers like /on, /id, /all etc. This function then returns a list of
matching tasks. This list can then be fed to another function mapTasks
that uses a consumer to execute the appropriate actions on each task.

This structure separates the main command, modifier command, and command
execution for all commands acting on tasks within the task list.
It makes it easier to extend functionality in each aspect.

Currently, the mark, unmark, delete and list commands support the
following modifiers:
- /all to execute for all tasks in the task list
- /on to execute by date
- /id to execute on the task with matching id
Adding a keyword without using any modifiers will execute the command
based on a keyword search.

Command classes have also been added for these commands to store
information to print. These are stored in the command package whereas
the LynxScanner and LynxCommandManager (used to be named LynxCommand)
have been moved to the parser package.