Create LynxCommand2 class and findTask2 method
(Test) To achieve more flexibility with search modifiers by
allowing chaining of search modifiers.

This would be done using a LynxCommand2 class whose purpose is to
break down a command string into parts and store them for use.

The findTask2 method would use a stream along with a LynxCommand2
object and recursively apply filter on the stream until no commands are
left to execute in the LynxCommand2.