Make LynxTaskList non abstract and make LynxSorter
To allow for a more object oriented design, LynxTaskList is now non
static, which allows for the use of multiple task lists at the same
time. Furthermore, the stream sorting methods were moved to a separate
abstract class, since they worked on streams and did not have to be tied
to any task list.