Improve general method naming and error handling logic
Some command method are vague and unintuitive, making the code
harder to understand and maintain. In addition, certain exception
handling blocks are inconsistent and do not provide clear error
messages.

This refactoring improves code readability and makes exception handling
more logical and helpful for users. Included mark and unmark command
in Storage as well for better coupling of logic.