Add ReminderCommand
Currently, the implementation of reminders are separate from the rest of
the commands in LynxTaskEditor.

However, the reminder is just a list command on a specific search string.

Therefore, a ReminderCommand class has been created with two factory
methods representing the two types of reminders.

This abstracts out the corresponding code from LynxTaskEditor.