Add ReminderCommand
The implementation of reminders are separate from the rest of
the commands in LynxTaskEditor.

However, the reminder is just a list command on a specified
search string.

Let's create a ReminderCommand class with two factory
methods representing the two types of reminders.

This abstracts out the corresponding code from LynxTaskEditor.