Extract task encoding in unloadTasks
The encoding of tasks for storage in a text data file is
carried out in the LynxStorage class.

The more types of classes there are, the more code needs to be added
in the unloadTasks method.

Let's make tasks themselves responsible for encoding their contents.
Implement a method storageRepresentation in the Task subclasses
for this purpose.

This follows an object oriented design and allows for flexibility
through polymorphism.