Extract task encoding in unloadTasks
Currently, the encoding of tasks for storage in a text data file is
carried out in the LynxStorage class. The downside is that the more
types of classes there are, the more code needs to be added in one
place.

To follow an object oriented design, tasks themselves are now
responsible for encoding their contents. Implement a method
storageRepresentation in the Task subclasses for this purpose.

This allows for more flexibility through polymorphism.