Add Assertions
Classes do not use assertions and rely on exception handling.

While exception handling is sufficient, it might not be
100% foolproof.

Thus I have added Assertions to help supplement the code.

Assertions are used to define assumptions about the program
state so that the runtime can verify them, thus are suitable
for making sure the code is running smoothly by checking
the invariants on the data at certain points.