Improve code readability and maintainability
Certain parts of Parser and storage are a little long and might not be easy to follow. There is also repition of code in TaskList.

By adding helper functions to Parser and storage, it is able to abstract out certain lines, making it easier to follow. Making a helper function for the repetitive in TaskList also helps with readability

Lets pull up behaviors common to the class into a helper function to increase readability.

Help ensure storage is clear and maintanable.