Add save functionality
App data does not persist across instances.

Storing user data on their device allows them to keep track of
their tasks even if they have to close the app (e.g. to shut down
their device).

Let's use Scanner and FileWriter to read and write to a text file,
storing the information of each task in a delimited format.

Using a text file to store data is simple and human editable,
which makes it easy to evolve and troubleshoot.