Add proper date/time format for data storage
Duke currently accepts any string as the input for dates in Deadline
and Event tasks, regardless of whether the string represents a date
in any way.

Change the type that we store it in from String to LocalDate, using
the relevant parsing and printing methods to ensure standardised
storage of info in the text file.

This change of type allows us to make use of the date/time of tasks
in meaningful ways in the future, for example, if we want to view the
tasks in order of due soonest to latest.