Organise the codebase using more OOP and packages
The codebase is currently disorganised, with most of the logic being
thrown into the main Clam.java file.

Organising the code based on OOP principles helps to compartmentalise
logic, creating a codebase that is easier to expand and evolve. Using
packages also helps us to keep files organised as well.

The enum variables have also been replaced with the Command class and
its respective subclasses, each corresponding to a possible command in
the program. Using an abstract Command class allows to generalise our
execution and allow polymorphism to ensure that the correct logic gets
run for each specific command.

This is done to ensure that our codebase is expandable as we add more
features and more lines of code.