Add find command
There is currently no way of searching or filtering tasks by name.

Being able to find tasks by searching will be useful for all users and
essential for power users who will have many tasks.

Let's add a find command that searches through all tasks in the task
list, and returns every task that contains the input string using regex.

This find command is case-sensitive. It is currently not of a high
priority to configure the find command to be case-insensitive, as the
necessity of it is minimal.