Add Assertions
There are currently no assertions in the code.

Assertions help us to verify that certain invariants are met throughout
our code.

Let's add assertions at points in our code where certain conditions
should be always true, but is not checked or guaranteed to be the case,
where these conditions are vital for the program's functionality.