Improve code quality of the codebase
The codebase has not been checked for code quality issues.

Ensuring a certain level of code quality is essential for readability,
which is especially important when maintaining a repo that other people
will be able to view, modify and update.

Let's review all existing code files and ensure that they follow the
code quality standards outlined in CS2103T's syllabus.