Add help command
There is currently no guide or help manual for users.

New users may not know what commands are available or what they do,
so having a help manual allows them to learn what these commands are to
be able to use Clam fully.

Let's add a detailed help message for every command available in Clam,
and have the help command itself print out all possible commands for
the user's reference. Additionally, if the program starts up and detects
that the task list is empty, it will assume that the user might be new
and show an additional message letting them know about the help command.