Add debug command
There is no debugging support built into the app.

The lack of debugging support within the app makes it slightly harder
for developers of the app to identify issues and bugs when testing the
app.

Let's add a "debug" command that prints out technical information such
as the java version it's being run on and the file path of the save
file.

This gives developers reliable information on the java version and file
path that that specific instance of the app is using, rather than having
to check through a separate terminal that might not be accurate.

The "debug" command is intentionally excluded from the list of commands
available in the "help" command, as it is not relevant for the average
user.