Update README.md
The README documentation is not updated to include the new "debug"
command.

The README documentation functions as a user guide for both new users
as well as developers intending to contribute/fork the repo. Providing
them with an exhaustive list of developed features gives them a good
overview of what is present in the codebase.

Let's add a header in the README.md to explain the "debug" command and
how to use it.