Bug fix: plurality of "task"/"tasks" in outputs
When running todo/event/deadline/delete commands, the output will
include a line that says "Now you have X tasks in the list." When X is
1, it still says "1 tasks", which is a grammatical error.

Although this bug does not affect the app's functionality, it is a QOL
(quality of life) issue that we would like to fix.

Let's update the relevant commands to check for the number of tasks in
the list when printing that line, and printing "task" instead of "tasks"
if there is only one task in the task list.

If there are no tasks in the task list, "tasks" is still the appropriate
form to use, so we only need to check for 1 or otherwise.