Chore: update javadoc
The javadoc for some of the newer java files are missing/partially
completed.

Javadoc is a useful tool for software development, allowing developers
to document how their methods and classes function in a way that is
useful for other coders to understand without having to analyse the code
itself. Since this codebase has already been consistently using javadoc
up until now, javadoc should be maintained for the newer classes and
methods for consistency and completion.

Let's add javadoc for the newer classes and methods that are missing
javadoc comments.